api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]

[tool.uv]
dev-dependencies = [
    "black>=25.1.0",
]
